Update README.md
AfriCanyon – Kloofing and Canyoning Adventure Activity in Plettenberg Bay
AfriCanyon in Plettenberg Bay was founded by Rogan Hindmarch. AfriCanyon was launched in order to share the multipart kloofing route, which lies in the hidden valleys of The Crags, near Plettenberg Bay. Whether you are a first-time adventurer or have loads of experience with abseiling. 
AfriCanyon - A brilliant combination of fun and adventure in the great outdoors! Book your AfriCanyon adventure with us!  Call 082 323 4349 or click http://www.africanyon.com/